+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 ## develop branch
 
-## New
+Nothing yet.
+
+## 2.5.1 - Wed Nov 2 2016
+
+### Fixes
+
+* dot.notation.path no longer has to contain a dot
+  - updated `IsDotNotationPath` check
+* the dot.notation.path classes now support appending to arrays on their own
+
+## 2.5.0 - Tue Feb 23 2016
+
+### New
 
 * Added interfaces to represent a read-write data container, and a read-only data container
   - Added `Containers\DataContainer` - read-write data container